fix: clear the post-rebase clippy gate (lints, merge-shape breakage, dead code)
No behavior change. Lint fixes across render/session/kuluu (unnecessary casts, collapsible match, struct literals over default-reassign, repeat_n, then_some, borrow_deref_ref), the parse_action_bytes 3-tuple and parse_action_bytes_reporting 4-tuple destructures, apply_action_dispatch's vestigial query params, the render_actor_with_movement_for_test helper restored to the new LoadedActor shape, and dead rebase leftovers removed (IMPACT_MIN_FRAME, assert_impact_at, AnchorStep, step_anchor).

+++ kuluu-render/src/cutscene_camera.rs
@@ -930,7 +930,7 @@ mod tests {
         let mut task = CutsceneCameraTask::start(
             &resource(
                 CameraSmoothType::Linear,
-                ffxi_dat::camera::CameraFlags::START_AT_CURRENT_POS as u16,
+                ffxi_dat::camera::CameraFlags::START_AT_CURRENT_POS,
                 vec![point([4.0, 0.0, 0.0], 350.0, [4.0, 1.0, 0.0], 0.0)],
             ),
             60.0,
@@ -951,7 +951,7 @@ mod tests {
         let mut task = CutsceneCameraTask::start(
             &resource(
                 CameraSmoothType::Linear,
-                ffxi_dat::camera::CameraFlags::END_AT_CURRENT_POS as u16,
+                ffxi_dat::camera::CameraFlags::END_AT_CURRENT_POS,
                 vec![point([0.0; 3], 280.0, [0.0, 1.0, 0.0], 0.0)],
             ),
             60.0,
@@ -1098,7 +1098,7 @@ mod tests {
         let mut task = CutsceneCameraTask::start(
             &resource(
                 CameraSmoothType::Linear,
-                ffxi_dat::camera::CameraFlags::START_AT_CURRENT_POS as u16,
+                ffxi_dat::camera::CameraFlags::START_AT_CURRENT_POS,
                 vec![point([0.0; 3], 280.0, [0.0; 3], 0.0)],
             ),
             60.0,
@@ -1284,8 +1284,8 @@ mod tests {
         // One authored point plus both flags is three effective points: a spline.
         let res = resource(
             CameraSmoothType::Linear,
-            (ffxi_dat::camera::CameraFlags::START_AT_CURRENT_POS
-                | ffxi_dat::camera::CameraFlags::END_AT_CURRENT_POS) as u16,
+            ffxi_dat::camera::CameraFlags::START_AT_CURRENT_POS
+                | ffxi_dat::camera::CameraFlags::END_AT_CURRENT_POS,
             vec![point([0.0; 3], 280.0, [0.0; 3], 0.0)],
         );
         assert_eq!(res.path_mode(), CameraPathMode::Spline);
